Set up BaseAppPerSuite
