+++ module/src/main/scala/org/scalatestplus/play/PlaySpec.scala
@@ -29,7 +29,7 @@ import play.api.http.{ HttpProtocol, Status, HeaderNames }
  * Convenience "super Suite" base class for Play tests.
  *
  * Extend this class by default for testing Play apps with the ScalaTest + Play library. You can mix other traits into it to access needed fixtures, such as
- * [[org.scalatestplus.play.OneAppPerSuite OneAppPerSuite]], [[org.scalatestplus.play.OneAppPerTest OneAppPerTest]], [[org.scalatestplus.play.OneServerPerSuite OneServerPerSuite]], [[org.scalatestplus.play.OneServerPerTest OneServerPerTest]], [[org.scalatestplus.play.OneBrowserPerSuite OneBrowserPerSuite]], [[org.scalatestplus.play.OneBrowserPerTest OneBrowserPerTest]], [[org.scalatestplus.play.AllBrowsersPerSuite AllBrowsersPerSuite]], or [[org.scalatestplus.play.AllBrowsersPerTest AllBrowsersPerTest]] mix If you want to use trait [[org.scalatestplus.play.MixedFixtures MixedFixtures]], extend [[org.scalatestplus.play.MixedPlaySpec MixedPlaySpec]] instead.
+ * [[org.scalatestplus.play.guice.GuiceOneAppPerSuite GuiceOneAppPerSuite]], [[org.scalatestplus.play.OneAppPerTest OneAppPerTest]], [[org.scalatestplus.play.OneServerPerSuite OneServerPerSuite]], [[org.scalatestplus.play.OneServerPerTest OneServerPerTest]], [[org.scalatestplus.play.OneBrowserPerSuite OneBrowserPerSuite]], [[org.scalatestplus.play.OneBrowserPerTest OneBrowserPerTest]], [[org.scalatestplus.play.AllBrowsersPerSuite AllBrowsersPerSuite]], or [[org.scalatestplus.play.AllBrowsersPerTest AllBrowsersPerTest]] mix If you want to use trait [[org.scalatestplus.play.MixedFixtures MixedFixtures]], extend [[org.scalatestplus.play.MixedPlaySpec MixedPlaySpec]] instead.
  */
 abstract class PlaySpec extends WordSpec with MustMatchers with OptionValues with WsScalaTestClient
