+++ module/src/test/scala/org/scalatestplus/play/examples/oneapppersuite/NestedExampleSpec.scala
@@ -18,7 +18,8 @@ package org.scalatestplus.play.examples.oneapppersuite
 import play.api.test._
 import org.scalatest._
 import org.scalatestplus.play._
-import play.api.{Play, Application}
+import org.scalatestplus.play.guice.GuiceOneAppPerSuite
+import play.api.{Application, Play}
 import play.api.inject.guice._
 
 // This is the "master" suite
@@ -27,9 +28,9 @@ class NestedExampleSpec extends Suites(
   new TwoSpec,
   new RedSpec,
   new BlueSpec
-) with OneAppPerSuite {
+) with GuiceOneAppPerSuite {
    // Override fakeApplication if you need an Application with other than non-default parameters.
-  def fakeApplication(): Application =
+  override def fakeApplication(): Application =
     new GuiceApplicationBuilder().configure(Map("ehcacheplugin" -> "disabled")).build()
 }
 
@@ -41,7 +42,7 @@ class NestedExampleSpec extends Suites(
 @DoNotDiscover
 class BlueSpec extends PlaySpec with ConfiguredApp {
 
-  "The OneAppPerSuite trait" must {
+  "The GuiceOneAppPerSuite trait" must {
     "provide an Application" in {
       app.configuration.getString("ehcacheplugin") mustBe Some("disabled")
     }
